+++ src/GUI.java
@@ -5,14 +5,14 @@ import java.awt.event.ActionListener;
 
 public class GUI extends JFrame {
 
-    Box box = Box.createVerticalBox();              // интерфейс на основе вертикального бокса  - сверху поле, снизу кнопки
-    JPanel jPanel = new JPanel();
-    static JTextField jTextField = new JTextField(60);
-    GridLayout gridLayout = new GridLayout(5, 4, 4, 2);
-    Font font = new Font("Verdana", Font.BOLD, 55);
+    private Box box = Box.createVerticalBox();  // интерфейс на основе вертикального бокса  - сверху поле, снизу кнопки
+    public static JTextField jTextField = new JTextField(60);      //здесь нужен public
+    private JPanel jPanel = new JPanel();
+    private GridLayout gridLayout = new GridLayout(5, 4, 4, 2);
+    private Font font = new Font("Verdana", Font.BOLD, 55);
 
-    String[] hints = new String[]{"1", "2", "3", "4", "5", "6", "7", "8", "9", "0", "+", "-", "*", "/", "="};      //подписи кнопок
-    JButton[] buttons = new JButton[15];
+    private  String[] hints = new String[]{"1", "2", "3", "4", "5", "6", "7", "8", "9", "0", "+", "-", "*", "/", "="};      //подписи кнопок
+    private  JButton[] buttons = new JButton[15];
 
     GUI(String title) {
         super(title);
@@ -31,7 +31,7 @@ public class GUI extends JFrame {
 
     }
 
-    public void sortButton() {                  // метод, добавляющий массив кнопок в интерфейс, закрепляя за ними слушателей
+    private void sortButton() {                  // метод, добавляющий массив кнопок в интерфейс, закрепляя за ними слушателей
 
         for (int i = 0; i < buttons.length; i++) {
             buttons[i] = new JButton(hints[i]);
@@ -39,7 +39,7 @@ public class GUI extends JFrame {
             if (i < 14)
                 buttons[i].addActionListener(new ButtonListener());
             else if (i==14)
-            buttons[14].addActionListener(new ResultListener());         // для знака равно другой слушатель
+                buttons[14].addActionListener(new ResultListener());         // для знака равно другой слушатель
 
         }
 
@@ -48,54 +48,10 @@ public class GUI extends JFrame {
     class ButtonListener implements ActionListener {                 // слушатель ввода. выводит текст из аргумента кнопки в поле
         @Override
         public void actionPerformed(ActionEvent event) {
-            String s = GUI.jTextField.getText();
-            GUI.jTextField.setText(s + event.getActionCommand());       // предыдущий текст + новый
-        }
-    }
-
-    class ResultListener implements ActionListener {           // слушатель вычисления после знака равно
-
-        @Override
-        public void actionPerformed(ActionEvent e) {
-            String first = "";                     //первая переменная
-            String second = "";                   //вторая переменная
-            String operation = "";                  //хранитель оператора вычисления
-            String[] strings = GUI.jTextField.getText().split("");       //превращаем текст из текстового поля в массив символов
-            Calculating calculating = new Calculating();                       //экземпляр вычисления
-
-            int i = 0;        //счетчик циклов. первый цикл принимает символы первого члена
-            while (!(strings[i].equals("+") || strings[i].equals("-") || strings[i].equals("*") || strings[i].equals("/"))) {
-                first += strings[i];
-                i++;
-            }
-              //условие работает на оператор, запоминая его в хранитель и записывая первый член в переменную x
-            if (strings[i].equals("+") || strings[i].equals("-") || strings[i].equals("*") || strings[i].equals("/")) {
-                calculating.x = Double.parseDouble(first);
-                operation = strings[i];
-                i++;
-            }
-
-            while (i != strings.length) {      //второй цикл для записи символов второго члена
-                second += strings[i];
-                i++;
-            }
-
-            calculating.y = Double.parseDouble(second);     //запись второго члена в переменную y
-
-            if (operation.equals("+")) {              // блок условий, определяющий, какой метод вычисления будет вызван. зависит от хранителя оператора вычисления
-                GUI.jTextField.setText(Double.toString(calculating.sum()));
-            } else if (operation.equals("-")) {
-                GUI.jTextField.setText(Double.toString(calculating.dif()));
-            } else if (operation.equals("*")) {
-                GUI.jTextField.setText(Double.toString(calculating.mult()));
-            } else if (operation.equals("/")) {
-                GUI.jTextField.setText(Double.toString(calculating.div()));
-            } else {
-                GUI.jTextField.setText("Кусь");
-            }
-
-
+            String previousText = GUI.jTextField.getText();
+            GUI.jTextField.setText(previousText + event.getActionCommand());       // предыдущий текст + новый
         }
     }
 
 }
+
